fix(micro-sam): float32 export image + wider export error tail (bump 0.4.2)
export_sam_model runs the model on the test image; SAM's apply_image_torch
interpolation is not implemented for uint8 ('Byte'), so cast the image to
float32. Widen the surfaced subprocess stderr (800->3000) to expose remaining
ModelDescr validation errors on the deployed micro-sam 1.8.5.

+++ apps/micro-sam/export_worker.py
@@ -27,7 +27,9 @@ def main(session_id: str, model_name: str) -> None:
     if not ckpt.exists():
         raise FileNotFoundError(f"No checkpoint for session '{session_id}'.")
 
-    image = tifffile.imread(p["train_images"][0])
+    # float32 image: export runs the model on this test image, and SAM's
+    # apply_image_torch interpolation is not implemented for uint8 ('Byte').
+    image = tifffile.imread(p["train_images"][0]).astype("float32")
     label = tifffile.imread(p["train_labels"][0])
 
     export_dir = training.session_dir(session_id) / "export"
